requirements.txt
mcp
python-dotenv
sentence-transformers
numpy
requests
